Lazily get explicit choice.
This fixes #2866 where with a very fresh project if you don't do a
BloopInstall and then you do a `generate-bsp-config` the explicit choice
gets inserted but the value we have is still the old one. This will
lazily get it to ensure we have the latest.

Note I didn't add an explicit test for this since we do a `bloopInstall`
right away in the build initialize in the Testing Server so it's hard to
actually test without making some large changes to that (which I'd like
to do in a separate pr).