Use workspace/didCreateFile for automatically filling package name
Since 3.16 (I think) there is an ability to register for FileCreated notifications, which we can use for PackageProvider. This is a better way of doing it, because we now get that information from the editor and not assume a file is new if it's empty.

For any client that doesn't support these notifications we keep the old way of adding the package.

Connected to https://github.com/scalameta/metals/issues/2869, but doesn't fix it unfortunately.