Disable Jmh and Scalafix build target
When importing Metals build with sbt as the build server, it creates:
- a bench-jmh build target, corresponding to the `bench / Jmh` scope
- many <project>-scalafix build targets, corresponding to the
`<project> / Scalafix` scopes.

Those are not very useful in the code editor because the user cannot
really modify there sources, and the `bench-jmh` build target is very
time consuming to import (it triggers the compilation of bench
to generate the sources).

By setting `bspEnabled := false` in those scopes, we remove those
build targets.