Wrap worksheets in object instead of a class
Previously, we would use the default mdoc modifier for worksheets, which uses a class to wrap statements. Now, we use `reset-object`, which wraps statements in an object and allows us to use some of the features not available in classes.

Fixes https://github.com/scalameta/metals/issues/2764 and https://github.com/scalameta/metals/issues/2907

Co-authored-by: Chris Kipp <ckipp@pm.me>