[Scala 3] Support extension methods in hover
Previously, hover would show extension methods as a simple method. Now, we try to show the extension signature next to the actual extension method signature. This isn't yet perfect as it will not work if a method has a using param as the first one. This is not a big issue though as parameter in both places will work exactly the same.