Customize showMessageRequest in TestClient
Responses of showMessageRequest used to be fixed in the TestingClient.
They are now mutable so that they can be customized on a test basis
For instance when asked whether to import an sbt build or not,
the client can answer yes to trigger a Bloop import or no and switch
to sbt as the build server.
(This is implemented in the subsequent commits)