Minimize file rewrites in `.metals/readonly/dependencies`
Previously, `toFileOnDisk` was always updating dependency source files in
order to keep them up to date.

After introducing `.metals/.readonly/dependencies/$jar-name` directories, it is possible to check if the source file is up to date and avoid non-necessary rewrites.
Now, this function also write `jar.meta` file that contains original jar path and modified time that is used for this purpose.

Co-authored-by: Tomasz Godzik <tgodzik@users.noreply.github.com>