[Scala3] Replace `InteractiveDriver` by our own wrapper to dotc.Compiler
`InteractiveDriver` was designed for dotty-language-server and we use it
improperly. For Metals we don't need to keep a buch of files in memory.
Also it might fix some problems(memory leaks + some pc tests) that might be caused of having a single
context that is passed trough all received sources.