Allow organize imports on unsaved code
Previously, if a file was not saved it wouldn't be possible to run the organize imports rule.

Now, we try to produce a new semanticdb, save it to disk and point scalafix to it. It might be better to transfer this directly, but scalafix is currently lacking an API for it. We can change once it's supported.

Fixes https://github.com/scalameta/metals/issues/2120