Handle Scala 3 style blocks of code without braces
Blocks with braces are always interpreted as longer ones. For example a block:
```
def foo =
  ???
  ???
```
is two lines long, while the same one, but with braces:
```
def foo = {
  ???
  ???
}
```
is four lines long.

Apart from this, the first line of the second block is the one containing `def`, while in the first
one it's the one with `???`.

Due to this, the line folding markers were incorrectly placed.

Fixes #3029