Fix issue with tmp directory popping up in random places
Previously, we would try to relatavize a path even if it isn't contained in the workspace. This might cause tmp folder with the created semanticdb popping up in weird places. The reason for that is that we would get a path like `../../../../tmp` and would resolve it from tmp directory. This sometimes would make the smeanticdb file generated for warming up scalafix to show in workspace directory.

Now, we only try to resolve files for organize imports for workspace files and also create the temporary warmup files in `.metals` directory