Incorporate show tasty into decode file
Fix bugs related to the package structure and multiple classes in same file by using quickpick. For now, if there are multiple classes for decode tasty or javap command, user had to choose one of them. If request is cancelled, both `value` and `error` are set to null.
Also, add tests for decode file scenarios