Wait for build tool information when generating semanticdb
Previously, we would wait until interactive sematnicdb was generate when opening a file, but this would cause an issue when a new workspace would be opened and we would wait for the semanticdb to be generated for files that belong to the project, which shouldn't be the case.
Now, interactive semanticdb will only be generated after connecting to a build tool, which will cause only the right files to have it generated. I also changed the synthetic provider to wait for both it and parsing, which is when we have the full information to show the synthetics.