Don't throw an exception is compiling semanticdb times out
Previously, we would throw an exception if we didn't manage to create a semanticdb file. This might happen when we are loading the project and at the same time try to create semanticdb for a dependency source. Now, we will return an empty and will only recompile after reopening.

Did should be retried after refocusing the document.