Wait with loading compiler and compilation
Previously, we would load the compiler as sonn as the workspace was opened, which would cause us to load a rambo compiler and print an invalid warning. We would also run the compilation, which possibly be skipped as the build tool was not loaded.

Now, we wait with compilation and loading the presentation compiler until we have information about the build tool (or lack thereof).