Add try in order not to break debugging session
While testing on windows I encounter still some issues for classes within `src.zip` JDK sources.

It was throwing FileSytemNotFound exceptions and I was not able to figure out what was going on.

This is a quickfix for this release so that it's possible to use breakpoints properly on Windows, the only thing this will do is that it might sometimes be impossible to go to JDK sources.