ScalaVersion - discover new mtags version by coursier
Prevously, the specific Metals version was able to work with only
specified list of Scala versions.

With these changes, in case if new Scala versions was released, we will
be able to realese only mtags artifact instead of making a full Metals
realease.

This isn't really usefull without additional release workflows but I'll
make separate PR for that later.

Co-authored-by: Kamil Podsiadło <37124721+kpodsiad@users.noreply.github.com>