docs: update commands that now use `TextDocumentPositionParams`
This was a small change that clients didn't account for in
https://github.com/scalameta/metals/commit/84dbb851ad753c6ac79a410a1b8aa272853db1f1
where the custom `GoToSuperMethodParams` were removed and just replaced
with `TextDocumentPositionParams`. This pr just updates the docs to
ensure that it's clear that the commands use this, not the old format.
Changes for this will need to happen in the client repos.