Fix expression for selection range in Ammonite
Ammonite generates additional file which we use in the presentation compiler for anything that requires types, so we neeed to adjust all request to actually point to a position in that generated file.

Previously, we would not adjust ranges when invoking hover and only positions. Now, we adjust both.

Also fixed the fallback version in case of Scala 3 for ammonite to be the last Ammonite suppored Scala 2.13 version instead of hte last Scala 2.13 version.