Simplify usage of sbt server when no .bsp/sbt.json exists
This relates to the conversation we had in
https://github.com/scalameta/metals/discussions/3255 and essentially
gets rid of the need for a user to remember they need to run
`metals.generate-bsp-config` if no `.bsp/sbt.json` exists. Now, instead
a user is able to use the `metals.bsp-switch` command and it will detect
if the build tool can be a build server even if there is no
`.bsp/<entry>` that exists, shows it to them during the user choice, and
then if selected generates the entry and then connects to it.

The old behavior of the `metals.generate-bsp-config` is still the same
and can be used _if_ you want a fresh `.bsp/<entry>`.