Move inactive maintainers to a separate list
There is a number of people, who haven't worked or interacted in any form with the Metals repository for over a year now. To avoid confusion and make it easier to know who can be pinged, I move them to a separate list.

I don't plan on modifying the permissions to go with it.