Use quick pick instead of message requests for debug discovery
Previously, we would use message request when asking for user input about which main class to run. Now, we instead use quick pick to accomodate a larger number of main classes possible. This is also a much nicer UI in VS Code as there is no need to select anything with a mouse. Quick pick still falls back to messageRequests if the client doesn't support it.