fix: skip extra search for main methods
Since we already have the mainClasses we can pass them right into
`verifyMain` just like we do when we use `run`. So this skips going
through the document symbols again and trying to get the main class. I'm
not 100% sure why this works for Scala 2 but not for Scala 2 using
`@main`, but this fixes it for me when using `@main`.

Closes #3458