docs: migrate gitter references to Discord
Following the official announcement that Discord is now where the
official Scala chat is located, I think it makes sense to focus on
support efforts on Discord. Lately I've also noticed that our Gitter has
been much quieter with more people coming to Discord anyways. We don't
have to sunset Gitter or anything, but instead just remove the
references to it in various places since most seem to be heading to
Discord anyways.