Add edit distance for Java files
Previously, we would not try to adjust positions in java files in case of modifications. Now, we use the Java tokenizer for that purpose.

I translate the Java and Scala tokens to custom BaseToken class, which can be used to find the proper edit distance.

Another issue that this cause is that it seems that MD5 string produced by the semanticdb for Java is different than the one we create in Metals. I have no idea why, but to accomodate it I copied over the Java implementation from sourcegraph/lsif-java and use it for Java files. That seems to have solved the issues.

This popped up while working on the rename for Java files.