fix: Open other window when using Analyze Stacktrace
When reworking the links and adding support for Sublime I forgot to test out if the stacktrace opens files in another window and it turns out this got broken.

Because of that, I reworked the GotoLocation command to take in an additional parameter, which makes it a bit easier to handle and always send it tpo the client.