Don't focus diagnotics if compialtion returned with error
Previously, we would focus diagnotics every time a compilation finished with an error, which means we could steal focus from the user multiple times. Now, the only palce we focus diagnostics is when running/debugging and the debug session failed to run due to compilation errors.

This has surfaced due to fixing an issue in the VS Code client.