docs(ammonite): add note about ammonite support
This comes from the comment in #2595. I add a small note about our
Ammonite not supporting multi-stage scripts and also not currrently
supporting Scala 3.

I also noticed we had a small "Unsupported features" section which was
out of date, but also very minimal. So I don't really think it's useful
unless we have a more exhaustive list of what is not supported.