Show implicit decorations in worksheets
Previously, we would not show implicit decorations in worksheet files, because they could get overriden by the worksheet decorations. Now, we send two different types of decorations, inline and not inline, which makes sure that one will not override the other.

This got tested in VS Code, but we would also need a confirmation on how to make it work in Sublime text.