feat, refactor: send test explorer updates to the clients (#3554)
* feat: send test explorer updates to the client
Previously, the client had to query the Metals server for all available test suites and the Metals server was responding with all of them every time. This functionality isn't removed, but it isn't default and encouraged way to interact with the Metals server.
Now, the Metals server will send updates (add a suite, remove a suite, add test cases) to the client on its own, using a push-based approach. Additionally, only relevant updates will be sent (not all of them like it was before) and this is the preferred way to use this feature.

* refactor: do not fetch semanticDB for every test suite
In https://github.com/scalameta/metals/pull/3605 symbol range was added to symbol definition. Symbol definition is created during the indexing and querying for the symbol definition is much faster than loading semanticDB from the file system.