Don't watch entire workspace on MacOS
Previously, on MacOS, the root of the workspace would be registered for
file watching. With this patch, up to 32 (by default) of the longest
prefixes will be registered.

This change improves the startup performance in large monorepos, where
the workspace is set to the root of the monorepo, but usually few
subdirectories are imported in the IDE.

The maximum number of roots to watch can be set to a different value
with the system propety `metals.macos-max-watch-roots`. Eg:

    -Dmetals.macos-max-watch-roots=128