fix: ensure user gets warned of workspace error when running with lens
Currently if the current workspace has an error but there is still a
code lens to run the project, we just forward the params and try to
start the debug server. This then fails since there are errors in the
workspace and doesn't really show anything to the user. This is just a
slight change that adds a check before that to see if there are
workspace errors and if so, show the warning.