feat: implement test selection request for sbt (#3678)
* feat: implement test selection request for sbt
Previously only bloop was capable of handling test selection suites request. Now, with sbt capable of handling it (https://github.com/scalacenter/scala-debug-adapter/pull/198), we have to change a few details in implementation. The most important ones are changes in "bsp" model which now are more aligned to the bsp spec and sbt's conraband capabilities.

* bump: scala debug adapter version
New version is needed because it implements abovementioned request