feat: add a statusbar message when starting debug server
Since there are times where this can take a bit to finish compiling your
project, this gives a bit more info into what's happening to the user
letting them know that the debug server is indeed starting, and then
once started a little message letting them know that.

This relates to #3730 but doesn't yet close it. Mainly because the
initialize itself can also take a few seconds, so I'd also like to get
some messaging there, but that one is a bit more involved. This one is
just a quick win.