docs: update vim docs to no longer point people to coc-metals
As of this next release
[coc-metals](https://github.com/scalameta/coc-metals) will be officially
deprecated, meaning we shouldn't be pointing people to it in the docs,
and also shouldn't have all the features outlined be coc specific. This
makes those changes.

Note that this is much much more minimal than it previously was. This is
mainly because the docs for `nvim-metals` are pretty spread out, and I
don't have any intention of duplicating them.

- The main docs are in
  https://github.com/scalameta/nvim-metals/blob/main/doc/metals.txt and
  they are in a specific format necessary for Vim help docs.
- https://github.com/scalameta/nvim-metals/discussions/39 holds the
  example configuration
- https://github.com/scalameta/nvim-metals/discussions/279 holds the
  feature list and instructions on how to utilize everything.

All this is mentioned and linked on the vim page.