test: Add MtagsScala3Suite
Introduce MtagsScala3Suite to check how mtags for Scala3 works.
MtagsScala3Suite doesn't compare the mtags-generated symbols and scalac-generated
symbols (which should be done in the future), and it doesn't check the
unkownSymbols, as we still have a lot to do in mtags for scala3.
It should be done in another PR.