feature: [Scala 3] Add completion item resolution
This code is mostly copied over from Scala 2 and adapted to Scala 3. I also moved a couple of parts to a common trait.

Now, in Scala 3 we get scaladoc and shows default values when moving accross completions. Previously, we would onlt show the ones from the compiler and eagerly, which is not needed.