fix: better override completions order
when we have the following methods in parent class

```scala
def hello1: Int
val hello2: Int
```

and metals oferring OverrideCompletion,
hello1 should be top over hello2.

However, before this change, hello2 will be top, because
hello1 is a field, and hello2 is not.

For override completions, we don't care it's field or non-field
because we can override both of them. That's why remove hasGetter
penalty for override-completion