Fix expression type for inlined methods and show all flags for methods
Previously, when an expression was inlined and was transparent we would not show the proper type as indicated by transparent. Besides that no flags were shown for methods. Now, we also check for Inlined tree node when expanding and also show the proper flag string.

Fixes https://github.com/scalameta/metals/issues/3929