feature: [Scala 3] Show correct inferred type in signature help
Previously, we would show the generic type of a method signature without trying to resolve the exact types. Now, we properly resolve types based on the qualifier.

One additional thing that we might improve is inferring type in case we already should be able to do so based on the arguments. For example:

```scala
fold("")(_ =>)
```

We already know that the first parameter will be a string, but we still show it as B.