fix: avoid out of bounds for editors that treat \n as a line end
So this isn't really a bug, but sort of an interested difference in how
different editors handle \n -- whether they treat it as a line
terminator or an actual new blank line when at the end of a file.

Currently when you use the new scala file provider the text you end up
getting is just the package followed by multiple new lines. Some editors
will interpret this as the package line and then two new blank lines,
while others (like vim) will treat this as a package line, one blank
line, and then a line ending. Then when the Metals calculates the
position of the cursor it will be out of bounds for vim. This small
change just ensures the `@@` for the new blank file template is not on
the last line meaning it won't be out of bounds. There will be a slight
difference in VS Code though as there will be an extra new line, but the
cursor won't be on that new line. I figured this was a fair compromise
to ensure it still worked on all editors. Is this alright?

refs: #3937