bugfix: Fallback to sourcepath if jar cannot be found
Previously, if something wasn't contained in sources set, we would assume that it's a dependency URI. IT seems however that in a rare scenario the set might not contain the specific source path while still being a valid workspace source.

Now, when that is the case we catch it in Try and fall back to just using it as a source.

I managed to reproduce it once, but not again. Not sure why the sources might be missing and it's probably safest to wrap it in Try