fix: lsif-java -> scip-java
The project was recently renamed to scip-java and lsif-java has been
[deprecated](https://github.com/coursier/apps/pull/153). I see we are
having an issue where no main class is being found, which was attempted
to be fixed [here](https://github.com/coursier/apps/pull/155), but I'm
still unable to run it locally. This just ensures that we are using the
most recent setup to ensure that our sourcegraph uploads are still
working.

For now, we'll use cs launch because the install is failing to pick up
the main method in scip-java