fix: only compare Bloop JavaHome with Metals JavaHome if set
Currently even if a user hasn't set the `javaHome` setting, but the
`.bloop/bloop.json` has a `javaHome` setting we try to compare them
which will always be false, since we're comparing `None !=
Some(javaHome)`. This changes the logic to compare the contained values
_only_ if they are set.

This is part of what is going on in #4001.