bugfix: Properly show links to implicit parameters in decoration hovers
Previously, no links would be shown for implicit parameter decorations in for comprehension. Now it's properly showed for the user.

The issue was that we were filtering synthatic decorations for a line based on the toplevel synthetic range, which in case of for comprehensions would be the full for comprehensions tree.

Fixes https://github.com/scalameta/metals/issues/3997