fix(worksheets): filter out NonUnitStatements wartremover in worksheets
I also hit on this during work and it annoys me. If you have a build
that is using wartremover and you are using the default set of
`Warts.unsafe` it includes the `NonUnitStatements` rule. When this rule
is part of your `scalacOptions` when you attempt to create a worksheet
next to your sources it will never work due to the binder in mdoc.
You'll get the following error at the top of your file and nothing will
evaluate:

```scala
val x = 3 // example piece of code in your worksheet
```
```
Diagnostics:
scala.worksheet.sc:9 (mdoc generated code) [wartremover:NonUnitStatements] Statements must return Unit
val x = 3; $doc.binder(x, 1, 4, 1, 5)
```

This change just strips out that rule so it doesn't get passed in during
evaluation so that your worksheet can actually evaluate.