feat(mill): also check default version in mill file
I just hit on this with https://github.com/com-lihaoyi/requests-scala
since they have a `mill` file which contains the default mill version
being used, but we only check for the version in the `.mill-version`
file. While most repos will have a `.mill-version` this adds an extra
fallback to check in the `mill` file to also grab it before we use the
default one we have. It can be the case that they are using an older
version of `mill` and our default version won't work to do the build
import. This guards against that.