feature: Don't update readonly bloop.json and allow to dismiss notification
Previously, metals would suggest to update bloop.json file even if it was readonly and there would be no way to dismiss the notification forever. Now, we will not suggest to update the file if it's readonly and users can ignore the notification if they want to. It's also possible to later reset all the notification so that everything can pop up again.