bugfix: Make sure that the Scala 3 compiler always retries properly
Previously, if the compiler was stuck in an infinite loop we would not be able to invoke any methods on the compiler. For Scala 2 we had a separate presentation compiler thread, which would get properly interupted stopped. Now, we need to reset the job queue. Testing locally this has helped. Unfortunately, I don't know how to reproduce the infinite loop and this should no longer be an issue for later Scala 3 versions

I also added handling of compiler exceptions, since there are currently none that I know of that should just we thrown and hnadled specifically.