feat: Do not provide convert-to-named-args code action in Scala3
The code action is not available in Scala3 project yet.
Previously, though we get the code action in Scala3 as if it works in
Scala3 projects, but if we run the action, we receive the error from
server.
It would be confusing for end users, so this commit hide the code action
in Scala3 project