bugfix: Allow completions in multiline expressions when debugging
Previously, I assumed that only single line is allowed when using expression evaluator, which turned out to be false. Now, we should provide proper completions even if the expression is multiline.

Unfortunately the logic is not as clear as I would like it to be, but making it work was not fun. The API for those completions is much worse than the normal one.