bugfix: Save fingerprints between restarts
Previously, if a user reloaded a broken workspace they would keep getting the dreaded "Could not load snapshot text for". Now, we save the fingerprints between restarts to improve the probability of finding the correct text, which corresponds to a semanicdb document.

It shouldn't save too much, since fingerprints are only updated for open files and we clean it after each reload.