feature: Add a way to turn on debug logging and fix scalafix warmup
Seems the main idea behind scribe is to make it easy to define in the code, so I decided to add an new metals property.

One issue is that Flyway debug logs still seem to end up in whichever handler is defined first, but this should not be an issue in most cases, since those logs are not super interesting.