bughack: Force specific Scala 3 compiler for worksheets
Mdoc is currently being released for Scala 3.1.x, which means that if we use mdoc for worksheets it will not be able to run with a classpath containing any Scala 3.2.0 code. However, since mdoc should be comptable with any future Scala 3 version, we can force a later compiler version to run with it.

This efectively fixes the worksheet issue for later Scala 3 versions. That is until the interfaces change, but hopefully we use a very limited api, so this should never be an issue.

Fixes https://github.com/scalameta/metals/issues/4138