chore(ci): migrate to actions/java to use built-in sbt caching
I'd like to see if this works. There were a couple coversations a while back about
using setup-scala vs setup-java and it seems that many repos have migrated to just
using setup-java since it does offer sbt support, has a ton of users, and also
built-in sbt caching which simplifies a few things in our actions.

If you're rather just leave it as is since it is working fine, no
worries, I just really like that the setup-java offers built-in caching
for sbt.